+++ config/test/webpack.test.js
@@ -6,7 +6,8 @@ const reactRule = {
   use: {
     loader: 'babel-loader',
     options: {
-      presets: [ '@babel/preset-react' ]
+      presets: [ '@babel/preset-react' ],
+      sourceMaps: true
     }
   }
 }