+++ src/AsyncReducerProvider.js
@@ -6,8 +6,10 @@ import { getReducerContext } from './ReducerProvider'
 export function AsyncReducerProvider(props) {
   const [ state, setter ] = React.useState(props.initialState)
   const wrappedDispatcher = React.useCallback(async (action) => {
-    setter(await props.reducer(state, action))
-  })
+    const nextState = await props.reducer(state, action)
+    setter(nextState)
+    return nextState
+  }, [ state ])
 
   return React.createElement(
     getReducerContext(props.name).Provider,